requirements-test.txt
shuup
mock
pytest>=3.6
pytest-cov
pytest-django>=3.3.2
pytest-sugar
flake8
isort
beautifulsoup4==4.5.3
faker==1.0.0
-e .
